prac_8/chat_service/requirements.txt
fastapi
uvicorn
sqlalchemy
alembic
psycopg2-binary
pydantic
python-multipart
minio
pika
python-dotenv
